+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # NEXT RELEASE
 * Add support for blocked domains
+* Precompile a mutedKeywordsRegex and store it in the UserData object
+* Add Toot.matchesRegex() method
 
 ### v0.69.22
 * Fix error handling in `MastoApi.getServerSideFilters()`